exercicios

+++ desafios/ex007.py
@@ -1,4 +1,4 @@
-nota1 = int(input('Digite a primeira nota: '))
-nota2 = int(input('Digite a segunda nota: '))
+nota1 = float(input('Digite a primeira nota: '))
+nota2 = float(input('Digite a segunda nota: '))
 media = (nota1 + nota2)/2
 print('A media da nota do aluno é {}'.format(media))
