+++ desafios/ex008.py
@@ -1,4 +1,4 @@
-medida = int(input('Digite quantos metros tem:'))
+medida = float(input('Digite quantos metros tem:'))
 cent = medida * 100
 milim = cent * 10
 print('{} metros tem {} centimetros e {} milimetros'.format(medida, cent, milim))
